reduce warnings

+++ ext/portvideo/interface/SDLinterface.cpp
@@ -164,12 +164,11 @@ void SDLinterface::updateDisplay()
         drawText(textHeight(),4*y,"U - cancel camera selection");
         drawText(textHeight(),5*y,"ENTER - apply camera selection");
 
-		
         CameraConfig *cfg = &dev_list[selector_];
-		char camera_str[256];
-		sprintf(camera_str,"%s%d: %s",dstr[cfg->driver],cfg->device,cfg->name);
-		drawText((width_- textWidth(camera_str))/2,height_/2-textHeight(),camera_str);
-		char format_str[256];
+	char camera_str[256];
+	sprintf(camera_str,"%s%d: %s",dstr[cfg->driver],cfg->device,cfg->name);
+	drawText((width_- textWidth(camera_str))/2,height_/2-textHeight(),camera_str);
+	char format_str[256];
         if (cfg->cam_fps==(int)cfg->cam_fps) sprintf(format_str,"%dx%d@%d (%s)",cfg->cam_width,cfg->cam_height,(int)cfg->cam_fps,fstr[cfg->cam_format]);
         else sprintf(format_str,"%dx%d@%.1f (%s)",cfg->cam_width,cfg->cam_height,cfg->cam_fps,fstr[cfg->cam_format]);
         drawText((width_- textWidth(format_str))/2,height_/2,format_str);
@@ -395,7 +394,7 @@ void SDLinterface::processEvents()
                 if (select_)
                 {
                     selector_++;
-                    if (selector_==dev_list.size()) selector_=0;
+                    if (selector_==(int)dev_list.size()) selector_=0;
                 }
             }
             else if( event.key.keysym.sym == SDLK_h )
